Changed Order model
Client:
- Widened the Register form on big screens
- Made necessary changes to api calls

Server:
- On order.model substituted property delivered for status
- In order.controller:
  -> Renamed and changed behviour of fetch function to get only closed orders
  -> Renamed and change behaviour of fetch-one function to get the open order
  -> Restricted behaviour of update function to only affect orders with status OPEN
  -> Added set Closed function to mark orders as CLOSED

+++ server/order/order.controller.js
@@ -1,16 +1,16 @@
-const { BAD_REQUEST, CREATED, NO_CONTENT, OK } = require('http-status-codes');
+const { BAD_REQUEST, CREATED, NO_CONTENT, NOT_MODIFIED, OK } = require('http-status-codes');
 const { Brand, Order, Product, ProductOrder } = require('../common/database');
 
-function fetch({ query: { userId } }, res) {
-  const where = userId ? { userId } : null;
+function getClosed({ query: { userId } }, res) {
+  const where = userId ? { userId, status: 'CLOSED' } : null;
   const include = includeAll();
   return Order.findAll({ where, include })
     .then(orders => res.send(orders))
     .catch(err => res.status(BAD_REQUEST).send(err));
 }
 
-function fetchOne({ query: { userId } }, res) {
-  const where = userId ? { userId } : null;
+function getOpen({ query: { userId } }, res) {
+  const where = userId ? { userId, status: 'OPEN' } : null;
   const include = includeAll();
   return Order.findOne({ where, include })
     .then(order => order ? res.send(order) : res.sendStatus(NO_CONTENT))
@@ -18,11 +18,11 @@ function fetchOne({ query: { userId } }, res) {
 }
 
 function create({ body: { userId, note, products } }, res) {
-  const where = userId ? { userId, delivered: false } : null;
+  const where = userId ? { userId, status: 'OPEN' } : null;
   return Order.findOne({ where })
     .then(openOrder => {
       if (openOrder) return res.sendStatus(208);
-      return Order.create({ delivered: false, userId, note })
+      return Order.create({ status: 'OPEN', userId, note })
         .then(order => processOrder(order, products))
         .then(() => res.sendStatus(CREATED));
     })
@@ -30,14 +30,26 @@ function create({ body: { userId, note, products } }, res) {
 }
 
 function update(req, res) {
-  const { params: { id }, body: { delivered, note, products } } = req;
+  const { params: { id }, body: { note, products } } = req;
   return Order.findByPk(id, { include: ProductOrder })
     .then(async order => {
-      await order.update({ delivered, note });
+      if (order.status !== 'OPEN') return res.sendStatus(NOT_MODIFIED);
+      await order.update({ note });
       await ProductOrder.destroy({ where: { orderId: id } });
       return processOrder(order, products);
     })
     .then(() => res.sendStatus(OK))
+    .catch(err => res.status(BAD_REQUEST).send(err));
+}
+
+function setClosed(req, res) {
+  const { params: { id }, body: { status } } = req;
+  return Order.findByPk(id)
+    .then(async order => {
+      if (order.status !== 'REVIEWED') return res.sendStatus(NOT_MODIFIED);
+      const done = await order.update({ status });
+      if (done) return res.sendStatus(OK);
+    })
     .catch(err => res.status(BAD_REQUEST).send(err));
 }
 
@@ -52,7 +64,7 @@ function remove({ params: { id } }, res) {
     .catch(err => res.status(BAD_REQUEST).send(err));
 }
 
-module.exports = { fetch, fetchOne, create, update, remove };
+module.exports = { getClosed, getOpen, create, update, setClosed, remove };
 
 async function processOrder(order, products) {
   const entries = products.map(({ productId, quantity }) => ({
@@ -63,6 +75,7 @@ async function processOrder(order, products) {
 }
 
 function includeAll() {
+  // TODO: Maybe include brandId, as well.
   const timestamps = [
     'createdAt',
     'updatedAt',
